requirements.txt
streamlit
python-pptx
python-dotenv
openai
pydub
moviepy

